python/pyproject.toml
[tool.poetry]
name = "jito_searcher_client"
version = "0.0.3"
description = "Jito Labs Python Searcher Client"
authors = ["Jito Labs <support@jito.wtf>"]
readme = "README.md"
packages = [{ include = "jito_searcher_client" }]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.51.1"
protobuf = "^4.21.12"
solders = "^0.14.2"
click = "^8.1.3"
solana = "^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-grpc-plugin]
proto_path = "../jito_protos/protos"
python_out = "./jito_searcher_client/generated"

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
